requirements.txt

streamlit==1.32.0
sentence-transformers==2.2.2
faiss-cpu
requests==2.31.0
PyPDF2==3.0.1
numpy==1.24.3
pandas==2.0.3

tavily-python>=0.1.0
openai>=0.27.0
requests>=2.28.0
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
